requirements.txt
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2022.5.18.1
cffi==1.15.0
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.2
defusedxml==0.7.1
Deprecated==1.2.13
Django==4.0.4
django-cors-headers==3.12.0
django-filter==21.1
django-stubs==1.12.0
django-stubs-ext==0.5.0
djangorestframework==3.13.1
djangorestframework-stubs==1.7.0
drf-spectacular==0.22.1
factory-boy==3.2.1
Faker==13.14.0
flake8==4.0.1
flake8-import-order==0.18.1
gunicorn==20.1.0
idna==3.3
importlib-resources==5.7.1
inflection==0.5.1
iniconfig==1.1.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.6.0
MarkupSafe==2.1.1
marshmallow==3.16.0
marshmallow-dataclass==8.5.8
mccabe==0.6.1
mypy==0.961
mypy-extensions==0.4.3
oauthlib==3.2.0
packaging==21.3
pluggy==1.0.0
psycopg2-binary==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.4.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
pytest-django==4.5.2
pytest-factoryboy==2.5.0
python-dateutil==2.8.2
python-dotenv==0.20.0
python3-openid==3.2.0
pytz==2022.1
PyYAML==6.0
redis==4.3.3
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.0.0
social-auth-core==4.2.0
sqlparse==0.4.2
tomli==2.0.1
types-Markdown==3.3.29
types-pytz==2022.1.1
types-PyYAML==6.0.9
types-redis==4.3.3
types-requests==2.28.0
types-urllib3==1.26.15
typing-inspect==0.7.1
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.9
wrapt==1.14.1
zipp==3.8.0
